Complete the basic framework of worker.
There are two basic functions of worker.
The first is parsing functions from message, and then listens on an address related to the functions selected by the user(programmer).
The second is to do the specified work when the user requests it.

The test program is placed in directory examples/worker_test.